+++ src/js/components/Route.js
@@ -1,9 +1,6 @@
 import Header from "./Header"
 import Footer from "./Footer"
-import TopPage from "./TopPage"
-import CommentBox from "./CommentBox"
-import AjaxCommentBox from "./AjaxCommentBox"
-import AjaxDiagnosis from "./AjaxDiagnosis"
+import Main from "./Main"
 import React from 'react';
 
 export default React.createClass ({
@@ -11,39 +8,10 @@ export default React.createClass ({
     return(
       <div>
         <Header/>
-        <Main/>
+        <Main url="http://localhost:3002"/>
         <Footer/>
       </div>
     );  // must be wrapped in an enclosing tag
   }
 })
 
-const Main = React.createClass ({
-  getInitialState : function(){
-    return {scene : "Top"}
-  },
-  changeScene(next_scene){
-    // this.state.scene = "Comment"     //← bad example. 
-    this.setState({"scene":next_scene}); //← good.
-  },
-  render() {
-    switch(this.state.scene){
-      case "Top":
-        return(
-          <TopPage changeScene={this.changeScene}/>
-        )
-      case "Comment":
-        return(
-          <CommentBox/>
-        )
-      case "AjaxComment":
-        return(
-          <AjaxCommentBox url="debug_commentbox"/>
-        )
-      case "AjaxDiagnosis":
-        return(
-          <AjaxDiagnosis interview_url="http://localhost:3002/api/interview" diagnosis_url="http://localhost:3002/api/diagnosis"/>
-        )
-    }
-  }
-})
