+++ testDiagnosisServer/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "testServer",
+  "name": "testDiagnosisServer",
   "version": "0.0.0",
   "scripts": {
     "start": "node ./"
